Improve text-based element finder to cover SVG elements
The "innertext" filter in elementfinder is based on elem.innerText, but SVG elements have no `.innerText` property, only `.textContent`. Now `.textContent` behaves differently on HTML nodes, and this modification doesn't want to introduce any breaking changes. It only "feature detects" `.innerText`, and if it's undefined, it falls back to `.textContent`. (`.innerText` is never undefined on HTML nodes.) The motivation of this PR comes from my experience not being able to click on highcharts chart slices, because of the aforementioned issue.